Update src/Symfony/Component/HttpKernel/Event/GetResponseForControllerResultEvent.php
Proposal: Adding a setControllerResult method to set manipulated controllerContent
There are for example  many listeners which convert the controller result 
to another format like XML or JSON. 
Without such a method i only can use the setContent method of Response
to define the new content. This content only can be a string and that's
a problem for me. The other listener wouldn't be able to manipulate a string
and convert it to another format. 